requirements.txt
# Core UI
streamlit==1.44.1
requests==2.32.5

# LLM Client
groq==0.30.0

# Data Handling
pandas==2.2.3

# Environment Variables
python-dotenv==1.1.0

# Vector Store for FAQ (ChromaDB)
chromadb==1.0.15

# Embeddings (used by ChromaDB & SemanticRouter)
sentence-transformers==5.0.0

# Semantic Routing
semantic-router[huggingface]==0.1.12

# FastAPI Backend
fastapi==0.115.5
uvicorn==0.32.1
slowapi==0.1.9
pydantic-settings==2.10.1
